margem_positiva(docker): adaptação para a utilização de docker e ajustes na compilação do código
