+++ README.md
@@ -1,1 +1,15 @@
 # Margem-Positiva
+
+Para rodar a aplicação, com o Docker ou Docker Desktop instalado, execute o seguinte comando:
+
+```shell
+docker-compose up --build
+```
+
+A aplicação automaticamente instalará as dependências e subirá os containers do aplicativo e do bando de dados
+
+Para encerrar a execução do contêiner utilize:
+
+```shell
+docker-compose down -v
+```
