+++ backend/docker-compose.yml
@@ -1,0 +1,39 @@
+version: '3.8'
+
+services:
+  app:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    container_name: margem-positiva
+    depends_on:
+      - db
+    ports:
+      - "8080:8080"
+    environment:
+      - SPRING_DATASOURCE_URL=jdbc:postgresql://db:5432/margem_positiva
+      - SPRING_DATASOURCE_USERNAME=margem
+      - SPRING_DATASOURCE_PASSWORD=margem
+    networks:
+      - net
+
+  db:
+    image: postgres:15-alpine
+    container_name: postgres15
+    environment:
+      - POSTGRES_DB=margem_positiva
+      - POSTGRES_USER=margem
+      - POSTGRES_PASSWORD=margem
+    ports:
+      - "5432:5432"
+    volumes:
+      - postgres_data:/var/lib/postgresql/data
+    networks:
+      - net
+
+networks:
+  net:
+    driver: bridge
+
+volumes:
+  postgres_data:
