+++ backend/src/main/java/br/com/margempositiva/backend/BackendApplication.java
@@ -1,0 +1,13 @@
+package br.com.margempositiva.backend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BackendApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BackendApplication.class, args);
+	}
+
+}
