+++ backend/src/main/resources/application.properties
@@ -1,1 +1,6 @@
 spring.application.name=backend
+
+server.port=8080
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
