--- backend/src/test/java/backend/BackendApplicationTests.java
@@ -1,13 +1,0 @@
-package backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
