+++ backend/src/test/java/br/com/margempositiva/backend/BackendApplicationTests.java
@@ -1,0 +1,13 @@
+package br.com.margempositiva.backend;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackendApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
